+++ ui/core/create-panels/advanced-options-panel.js
@@ -17,7 +17,7 @@ import { Audio } from '/core/ui/audio-base/audio-support.js';
 const STANDARD_PARAMETERS = ["Age", "Difficulty", "GameSpeeds", "Map", "MapSize"];
 const VICTORY_PARAMETERS = ["MilitaryVictoryEnabled", "ScienceVictoryEnabled", "EconomicVictoryEnabled", "CultureVictoryEnabled"];
 const ADVANCED_PARAMETERS = ["AgeLength", "DisasterIntensity", "CrisesEnabled", "GameRandomSeed", "MapRandomSeed", "StartPosition"];
-const PROGRESSION_PARAMETERS = ["AgeProgressionFromPlayerEliminatedEnabled"];
+const PROGRESSION_PARAMETERS = ["AgeProgressionFromTurnCounterEnabled", "AgeProgressionFromPlayerEliminatedEnabled"];
 
 /**
  * AdvancedOptionsPanel displays advanced game options and player setup.